Rails 6 Upgrade
Merged in staging (pull request #996)

+++ lib/groovepacker/stores/importers/importer.rb
@@ -56,14 +56,14 @@ module Groovepacker
 
         def initialize_import_item
           total_imported = @result[:total_imported] || 1
-          @import_item.update_attributes(current_increment_id: '',
-                                         success_imported: 0,
-                                         previous_imported: 0,
-                                         updated_orders_import: 0,
-                                         current_order_items: -1,
-                                         current_order_imported_item: -1,
-                                         to_import: total_imported)
-          sleep 0.5
+          @import_item.update(current_increment_id: '',
+                              success_imported: 0,
+                              previous_imported: 0,
+                              updated_orders_import: 0,
+                              current_order_items: -1,
+                              current_order_imported_item: -1,
+                              to_import: total_imported)
+          sleep 0.5 unless Rails.env.test?
           import_item_fix
           import_summary = OrderImportSummary.top_summary
           import_summary&.emit_data_to_user(true)
@@ -72,20 +72,19 @@ module Groovepacker
         def import_item_fix
           new_import_item = @import_item
           @import_item = begin
-                           ImportItem.find(@import_item.id)
-                         rescue StandardError
-                           new_import_item
-                         end
+            ImportItem.find(@import_item.id)
+          rescue StandardError
+            new_import_item
+          end
         end
 
         def fix_import_item(import_item)
           new_import_item = import_item
-          import_item = begin
-                          ImportItem.find(import_item.id)
-                        rescue StandardError
-                          new_import_item
-                        end
-          import_item
+          begin
+            ImportItem.find(import_item.id)
+          rescue StandardError
+            new_import_item
+          end
         end
 
         def update_success_import_count
@@ -125,20 +124,18 @@ module Groovepacker
         def handle_cancelled_order(gp_order)
           return false unless gp_order.present? && gp_order.status == 'cancelled'
 
-          if @credential.remove_cancelled_orders
-            gp_order.destroy
-          end
+          gp_order.destroy if @credential.remove_cancelled_orders
           true
         end
 
         def ondemand_user(user_id)
-          return unless user_id 
+          return unless user_id
           user_name = User.find_by_id(user_id).name
           "(#{user_name})" if user_name
         end
 
         def check_or_assign_import_item
-          return unless ImportItem.find_by_id(@import_item.id).blank?
+          return if ImportItem.find_by_id(@import_item.id).present?
 
           import_item_id = @import_item.id
           @import_item = @import_item.dup
@@ -147,13 +144,15 @@ module Groovepacker
         end
 
         def update_orders_status
-          result = { 'status' => true, 'messages' => [], 'error_messages' => [], 'success_messages' => [], 'notice_messages' => [] }
-          Groovepacker::Orders::BulkActions.new.delay(priority: 95).update_bulk_orders_status(result, {}, Apartment::Tenant.current)
+          result = { 'status' => true, 'messages' => [], 'error_messages' => [], 'success_messages' => [],
+                     'notice_messages' => [] }
+          Groovepacker::Orders::BulkActions.new.delay(priority: 95).update_bulk_orders_status(result, {},
+                                                                                              Apartment::Tenant.current)
         end
 
         def log_import_error(e)
-          on_demand_logger = Logger.new("#{Rails.root}/log/import_error_logs.log")
-          log = { time: Time.zone.now, tenant: Apartment::Tenant.current, e: e, backtrace: e.backtrace.join(',') }
+          on_demand_logger = Logger.new("#{Rails.root.join('log/import_error_logs.log')}")
+          log = { time: Time.zone.now, tenant: Apartment::Tenant.current, e:, backtrace: e.backtrace.join(',') }
           on_demand_logger.info(log)
         rescue StandardError
         end
@@ -170,13 +169,17 @@ module Groovepacker
             order_in_gp_present = true
             is_scanned = order_in_gp && (order_in_gp.status == 'scanned' || order_in_gp.status == 'cancelled' || order_in_gp.order_items.map(&:scanned_status).include?('partially_scanned') || order_in_gp.order_items.map(&:scanned_status).include?('scanned'))
             # mark previously imported
-            update_import_count('success_updated') && return if is_scanned || (order_in_gp.last_modified == Time.zone.parse(order['updated_at']))
+            if is_scanned || (order_in_gp.last_modified == Time.zone.parse(order['updated_at']))
+              update_import_count('success_updated') && return
+            end
+
             order_in_gp.order_items.destroy_all
           else
-            order_in_gp = Order.new(increment_id: order['number'], store: @store, store_order_id: order['id'].to_s, importer_id: @worker_id, import_item_id: @import_item.id)
+            order_in_gp = Order.new(increment_id: order['number'], store: @store, store_order_id: order['id'].to_s,
+                                    importer_id: @worker_id, import_item_id: @import_item.id)
           end
           import_order_and_items(order, order_in_gp)
-          order_in_gp.add_tag_from_shopify(order['tags']) if order['tags'].present? && @store.store_type == "Shopify"
+          order_in_gp.add_tag_from_shopify(order['tags']) if order['tags'].present? && @store.store_type == 'Shopify'
 
           # increase successful import with 1 and save
           order_in_gp_present ? update_import_count('success_updated') : update_import_count('success_imported')
@@ -186,17 +189,17 @@ module Groovepacker
             log_import_error(e)
           rescue StandardError
             nil
-            end
+          end
           update_import_count('success_imported')
         end
 
         def update_import_count(import_type = 'success_imported')
           if import_type == 'success_imported'
-            @import_item.update_attributes(success_imported: @import_item.success_imported + 1)
+            @import_item.update(success_imported: @import_item.success_imported + 1)
             @result[:success_imported] += 1
           else
             @result[:previous_imported] += 1
-            @import_item.update_attributes(updated_orders_import: @import_item.updated_orders_import + 1)
+            @import_item.update(updated_orders_import: @import_item.updated_orders_import + 1)
           end
         end
